feat: TUI のフォワード行にルール名を表示する (#64)
ForwardRow.View() でホスト名の後、転送種別の前にルール名を太字で描画。
狭いターミナル幅では名前を動的にトランケートする。
テストをforwardrow_test.goに分離しファイル行数制限に対応。
